requirements.txt
streamlit==1.24.1
torch==2.3.1
llama-index>=0.8.5
langchain-community
ibm-watson-machine-learning>=1.0.320
vertexai
google-cloud-aiplatform>=1.38
python-dotenv==1.0.0
scikit-learn
pandas
flask
openai
psycopg2-binary
pgcli
pillow
gunicorn